formats as list-classes
formats are now child-classes of list. The constructors took a lot of
arguments to control which formats should be built. If a string is given
it will be evaluated and the generation of formats will be tailored to
it.

Another concept of these classes is that they weet out a existing list
of formats according to a string. But this isn't yet realized straigth.
Maybe I'll skip that feature...