ENDIAN- and TODAY-classes
ENDIAN is now an iterable object with a set-method to change the
order-tuple to iter over. set accepts an string as argument which is
checked for matching an endian-key (little, big or medium).

TODAY is a datetime.date that defaults to datetime.date.today(), which
also has a set-method to change the date of TODAY. Tooks a
datetime.date-instance as argument.

TODO: write docstrings...